cloud-run/requirements.txt
flask==3.0.0
gunicorn==21.2.0
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.18.0
google-cloud-pubsub==2.19.0
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-generativeai>=0.8.0
pdf2image==1.17.0
Pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.1
google-cloud-logging==3.9.0
